refactor(linter): extract definitely_returns_in_all_codepaths into shared utils
